Revert "[CI] Split test_multi_node to avoid timeouts (#13712)"
This reverts commit cf2ffa7941a19acf4eaec78a2199636f0868fead.
